Level2 Association naming complete
WIP #639
