+++ jwst/associations/lib/dms_base.py
@@ -1,0 +1,49 @@
+"""Association attributes common to DMS-based Rules"""
+
+from jwst.associations.lib.acid import ACIDMixin
+
+# DMS file name templates
+_ASN_NAME_TEMPLATE_STAMP = 'jw{program}-{acid}_{stamp}_{type}_{sequence:03d}_asn'
+_ASN_NAME_TEMPLATE = 'jw{program}-{acid}_{type}_{sequence:03d}_asn'
+
+__all__ = ['DMSBaseMixin']
+
+
+class DMSBaseMixin(ACIDMixin):
+    """Association attributes common to DMS-based Rules"""
+
+    @property
+    def acid(self):
+        """Association ID"""
+        return self.acid_from_constraints()
+
+    @property
+    def asn_name(self):
+        program = self.data['program']
+        version_id = self.version_id
+        asn_type = self.data['asn_type']
+        sequence = self.sequence
+
+        if version_id:
+            name = _ASN_NAME_TEMPLATE_STAMP.format(
+                program=program,
+                acid=self.acid.id,
+                stamp=version_id,
+                type=asn_type,
+                sequence=sequence,
+            )
+        else:
+            name = _ASN_NAME_TEMPLATE.format(
+                program=program,
+                acid=self.acid.id,
+                type=asn_type,
+                sequence=sequence,
+            )
+        return name.lower()
+
+    def __ne__(self, other):
+        """Compare inequality of two associations"""
+        if isinstance(other, self.__class__):
+            return not self.__eq__(other)
+        else:
+            return NotImplemented
