+++ jwst/associations/tests/test_asn_names.py
@@ -11,7 +11,7 @@ LEVEL3_ASN_ACID_NAME_REGEX = (
     'jw'
     '(?P<program>\d{5})'
     '-(?P<acid>(o|c)\d{3,4})'
-    '_(?P<asn_type>[a-z]+)'
+    '_(?P<asn_type>\w+)'
     '_(?P<sequence>\d{3})'
     '_asn'
 )
@@ -19,7 +19,7 @@ LEVEL3_ASN_DISCOVERED_NAME_REGEX = (
     'jw'
     '(?P<program>\d{5})'
     '-(?P<acid>a\d{4})'
-    '_(?P<asn_type>[a-z]+)'
+    '_(?P<asn_type>\w+)'
     '_(?P<sequence>\d{3})'
     '_asn'
 )
@@ -43,32 +43,58 @@ pool_params = pytest.fixture(
 )(helpers.generate_params)
 
 
-class TestASNtNames():
+def test_level3_asn_names(pool_params):
+    pool_path = helpers.t_path(pool_params)
+    pool = helpers.combine_pools(pool_path)
+    rules = helpers.registry_level3_only()
+    asns, orphaned = generate(pool, rules)
+    assert len(asns) > 0
+    for asn in asns:
+        name = asn.asn_name
+        if any(
+                c.get('is_acid', False)
+                for _, c in asn.constraints.items()
+        ):
+            m = re.match(LEVEL3_ASN_ACID_NAME_REGEX, name)
+        else:
+            m = re.match(LEVEL3_ASN_DISCOVERED_NAME_REGEX, name)
+        assert m is not None
 
-    def test_level3_asn_names(self, pool_params):
-        pool_path = helpers.t_path(pool_params)
-        pool = helpers.combine_pools(pool_path)
-        rules = helpers.registry_level3_only()
-        asns, orphaned = generate(pool, rules)
-        assert len(asns) > 0
-        for asn in asns:
-            name = asn.asn_name
-            if any(
-                    c.get('is_acid', False)
-                    for _, c in asn.constraints.items()
-            ):
-                m = re.match(LEVEL3_ASN_ACID_NAME_REGEX, name)
-            else:
-                m = re.match(LEVEL3_ASN_DISCOVERED_NAME_REGEX, name)
-            assert m is not None
+def test_level3_asn_names_with_version(pool_params):
+    pool_path = helpers.t_path(pool_params)
+    pool = helpers.combine_pools(pool_path)
+    rules = helpers.registry_level3_only()
+    asns, orphaned = generate(pool, rules, version_id=True)
+    assert len(asns) > 0
+    for asn in asns:
+        name = asn.asn_name
+        m = re.match(LEVEL3_ASN_WITH_VERSION, name)
+        assert m is not None
 
-    def test_level3_asn_names_with_version(self, pool_params):
-        pool_path = helpers.t_path(pool_params)
-        pool = helpers.combine_pools(pool_path)
-        rules = helpers.registry_level3_only()
-        asns, orphaned = generate(pool, rules, version_id=True)
-        assert len(asns) > 0
-        for asn in asns:
-            name = asn.asn_name
-            m = re.match(LEVEL3_ASN_WITH_VERSION, name)
-            assert m is not None
+def test_level2_asn_names(pool_params):
+    pool_path = helpers.t_path(pool_params)
+    pool = helpers.combine_pools(pool_path)
+    rules = helpers.registry_level2_only()
+    asns, orphaned = generate(pool, rules)
+    assert len(asns) > 0
+    for asn in asns:
+        name = asn.asn_name
+        if any(
+                c.get('is_acid', False)
+                for _, c in asn.constraints.items()
+        ):
+            m = re.match(LEVEL3_ASN_ACID_NAME_REGEX, name)
+        else:
+            m = re.match(LEVEL3_ASN_DISCOVERED_NAME_REGEX, name)
+        assert m is not None
+
+def test_level2_asn_names_with_version(pool_params):
+    pool_path = helpers.t_path(pool_params)
+    pool = helpers.combine_pools(pool_path)
+    rules = helpers.registry_level2_only()
+    asns, orphaned = generate(pool, rules, version_id=True)
+    assert len(asns) > 0
+    for asn in asns:
+        name = asn.asn_name
+        m = re.match(LEVEL3_ASN_WITH_VERSION, name)
+        assert m is not None
